requirements.txt
python-miio>=0.5.12
homeassistant>=2023.11.3
voluptuous>=0.14.1